+++ src/Functions.Worker.Extensions.OpenAI/Search/ConnectionInfo.cs
@@ -8,12 +8,14 @@ using System.Text;
 namespace Microsoft.Azure.Functions.Worker.Extensions.OpenAI.Search;
 public class ConnectionInfo
 {
-    public ConnectionInfo(string ConnectionName, string CollectionName)
+    public ConnectionInfo(string ConnectionName, string CollectionName, string? Credentials)
     {
         this.ConnectionName = ConnectionName;
         this.CollectionName = CollectionName;
+        this.Credentials = Credentials;
     }
 
     public string ConnectionName { get; set; }
     public string CollectionName { get; set; }
+    public string? Credentials { get; set; }
 }